requirements.txt
# Core dependencies
openai>=1.58.1
tiktoken>=0.8.0
pyyaml>=6.0
pydantic>=2.10.4
requests>=2.31.0

# Database
sqlalchemy>=2.0.36
alembic>=1.13.0

# File processing
PyPDF2>=3.0.1
jsonschema>=4.20.0

# Utilities
tenacity>=9.0.0
python-dotenv>=1.0.1

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
hypothesis>=6.82.0
